requirements.txt
#requirements.txt
flask
pandas
openpyxl
waitress
python-dotenv